Fix overflow reporting bug
In 2023d36 Bugbear was complaining that the overflow variable
was not used in the loop. By mistake I deleted the variable
without realizing that past me was clever enough to be doing
the right thing -- we only need the value of "overflow" during
the final go round the loop. Rearrange to be more explicit
even though this is less efficient.